+++ frontend/src/components/Input.jsx
@@ -1,7 +1,8 @@
 import { useRef, useState, useEffect } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faSearch, faMapLocationDot, faHandPointer, faThumbsUp } from '@fortawesome/free-solid-svg-icons';
-import { Link } from 'react-router-dom';
+import { faSearch, faHandPointer } from '@fortawesome/free-solid-svg-icons';
+import Modal from './Modal';
+import PropTypes from 'prop-types';
 
 const Input = ({title}) => {
   const [formVisible, setFormVisible] = useState(false);
@@ -10,7 +11,6 @@ const Input = ({title}) => {
   const [results, setResults] = useState([]); 
   const [modalVisible, setModalVisible] = useState(false); 
   const [selectedResult, setSelectedResult] = useState(null);
-  const [copiedSpan, setCopiedSpan] = useState(null);
   const inputRef = useRef(null);
   const buttonRef = useRef(null);
   const containerRef = useRef(null);
@@ -50,7 +50,6 @@ const Input = ({title}) => {
       setIsLoading(true);
       setResults([]);
       try {
-        // const tempUrl = `http://localhost:8080/api/v1/search/api-point-with-page?query=${addressValue}&page=1`;
         const tempUrl = `/api/v1/search/api-point-with-page?query=${addressValue}&page=1`;
         const response = await fetch(tempUrl);
         const data = await response.json();
@@ -99,25 +98,7 @@ const Input = ({title}) => {
     }
   };
 
-  const copyToClipboard = async (text) => {
-    try {
-      await navigator.clipboard.writeText(text);
-      console.log('Text copied to clipboard');
-    } catch (err) {
-      console.error('Error copying text: ', err);
-    }
-  };
 
-  const handleCopyClick = async (e, value) => {
-    if (copiedSpan !== null) {
-      return;
-    }
-    setCopiedSpan(value);
-    await copyToClipboard(value);
-    setTimeout(() => {
-      setCopiedSpan(null);
-    }, 1000);
-  };
 
   return (
     <>
@@ -168,69 +149,22 @@ const Input = ({title}) => {
                     )}
                   </div>
                 ))}
-
               </ul>
             </div>
           )}
         </div>
       </div>
-
-      <div onClick={closeModal} className={`modal-window ${modalVisible ? 'modal-visible' : ''}`}>
-        {selectedResult && (
-          <div>
-            <div onClick={closeModal} className="modal-close">Close</div>
-            <div className='address'>
-              <span
-                className='copy-span' 
-                onClick={(e) => handleCopyClick(e, selectedResult.roadNameAddress)}
-              >
-                {copiedSpan === selectedResult.roadNameAddress ? <span className='copied-span'>copy <FontAwesomeIcon icon={faThumbsUp} /></span> : selectedResult.roadNameAddress}
-              </span>
-            </div>
-            <div>
-              <span
-                className='copy-span' 
-                onClick={(e) => handleCopyClick(e, selectedResult.parcelAddress)}
-              >
-                {copiedSpan === selectedResult.parcelAddress ? <span className='copied-span'>copy <FontAwesomeIcon icon={faThumbsUp} /></span> : selectedResult.parcelAddress}
-              </span>
-            </div>
-            <div>우편 번호 :&nbsp;
-              <span
-                className='copy-span' 
-                onClick={(e) => handleCopyClick(e, selectedResult.zipcode)}
-              >
-                {copiedSpan === selectedResult.zipcode ? <span className='copied-span'>copy <FontAwesomeIcon icon={faThumbsUp} /></span> : selectedResult.zipcode}
-              </span>
-              </div>
-            <div>
-              위도 :&nbsp;
-              <span 
-                className='copy-span' 
-                onClick={(e) => handleCopyClick(e, selectedResult.x)}
-              >
-                {copiedSpan === selectedResult.x ? <span className='copied-span'>copy <FontAwesomeIcon icon={faThumbsUp} /></span> : selectedResult.x}
-              </span>
-            </div>
-            <div>
-              경도 :&nbsp;
-              <span 
-                className='copy-span' 
-                onClick={(e) => handleCopyClick(e, selectedResult.y)}
-              >
-                {copiedSpan === selectedResult.y ? <span className='copied-span'>copy <FontAwesomeIcon icon={faThumbsUp} /></span> : selectedResult.y}
-              </span>
-            </div>
-            <Link to={`/coordinate/map?x=${selectedResult.x}&y=${selectedResult.y}`}>
-              <div className='map-Link'>
-                <FontAwesomeIcon icon={faMapLocationDot} /> 위치 보기
-              </div>
-            </Link>
-          </div>
-        )}
-      </div>
+      <Modal 
+        modalVisible={modalVisible}
+        selectedResult={selectedResult}
+        closeModal={closeModal}
+      />
     </>
   );
 };
+
+Input.propTypes = {
+  title: PropTypes.string
+}
 
 export default Input;
